requirements.txt
# FastAPI and web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Database drivers
motor>=3.3.2  # MongoDB async driver
redis>=5.0.1
pymongo>=4.6.1

# Vector databases (choose one or install both)
# chromadb>=0.4.22  # Uncomment if using Chroma
qdrant-client>=1.7.0  # Uncomment if using Qdrant

# Scheduling
apscheduler>=3.10.4

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.9
pyyaml>=6.0.1  # For simulator config
